// src/web/shared/proto/jaiabot/messages/mission.ts
export enum MissionTask_TaskType {
    NONE = "NONE",
    DIVE = "DIVE",
    STATION_KEEP = "STATION_KEEP",
    SURFACE_DRIFT = "SURFACE_DRIFT",
    CONSTANT_HEADING = "CONSTANT_HEADING",
}